chore: inputWrapper 컴포넌트 삭제
