+++ packages/ui/src/imageUploader/ImageUploader.tsx
@@ -10,27 +10,26 @@ const ImageUploader = () => {
   const [selectItem, setSelectItem] = useState<UrlType>();
 
   const handleDrag = (e: React.DragEvent, urlItem: UrlType) => {
-    //드래그 시작할때
+    //프리뷰 아이템 드래그 시작할때
     e.stopPropagation();
     setSelectItem(urlItem);
   };
-  const handleDrop = (e: React.DragEvent, targetItem: UrlType) => {
+  const handleDrop = (e: React.DragEvent, targetIndex: number) => {
     // 드롭 시, urlList에서 순서를 바꿔주기
     e.preventDefault();
     e.stopPropagation();
-    if (selectItem) {
-      const prevList = [...urlList];
-      // 드래그한 항목의 인덱스 찾기
-      const fromIndex = prevList.indexOf(selectItem);
-      const toIndex = prevList.indexOf(targetItem);
+    if (!selectItem) return;
 
-      // 순서 바꾸기
-      if (fromIndex !== -1 && toIndex !== -1 && fromIndex !== toIndex) {
-        // 배열의 순서를 바꿔줍니다.
-        prevList.splice(fromIndex, 1); // fromIndex 항목을 삭제
-        prevList.splice(toIndex, 0, selectItem); // toIndex 위치에 selectItem 삽입
-        setUrlList(prevList); // 업데이트된 리스트 상태로 설정
-      }
+    const prevList = [...urlList];
+    // 드래그한 항목의 인덱스 찾기
+    const fromIndex = prevList.indexOf(selectItem);
+
+    // 순서 바꾸기
+    if (fromIndex !== -1 && targetIndex !== -1 && fromIndex !== targetIndex) {
+      // 배열의 순서를 바꿔줍니다.
+      prevList.splice(fromIndex, 1); // fromIndex 항목을 삭제
+      prevList.splice(targetIndex, 0, selectItem); // toIndex 위치에 selectItem 삽입
+      setUrlList(prevList); // 업데이트된 리스트 상태로 설정
     }
   };
   const handleDragEnter = (e: React.DragEvent) => {
@@ -42,11 +41,11 @@ const ImageUploader = () => {
     <div className={style.wrapper}>
       <ImageInput />
       <ul className={style.previewList}>
-        {urlList.map((urlItem) => (
+        {urlList.map((urlItem, index) => (
           <li
             key={urlItem.id}
             draggable
-            onDrop={(e) => handleDrop(e, urlItem)}
+            onDrop={(e) => handleDrop(e, index)}
             onDragOver={handleDragEnter}
             onDragStart={(e) => handleDrag(e, urlItem)}
           >
