+++ packages/ui/src/imageUploader/components/imageInput/style.css.ts
@@ -29,3 +29,14 @@ export const label = styleVariants({
   ],
   none: [base],
 });
+
+export const wrapper = style({
+  position: "relative",
+  border: "1px solid",
+  borderColor: Color.border,
+  borderRadius: "8px",
+  padding: "8px",
+  display: "flex",
+  backgroundColor: Color.primary.default,
+  cursor: "pointer",
+});
